chore: modify cwd of test cases and refactor some case creator
