+++ packages/rspack-test-tools/src/helper/util/currentWatchStep.js
