+++ packages/rspack-test-tools/tests/builtinCases/plugin-css-modules/modules-composes/test.config.js
@@ -1,4 +1,4 @@
-/** @type {import("../../../../../rspack").Configuration} */
+/** @type {import("@rspack/core").Configuration} */
 module.exports = {
 	module: {
 		rules: [
