+++ packages/rspack-test-tools/tests/fixtures/hook.js
@@ -1,0 +1,1 @@
+module.exports = "This is hook"
